+++ python_modules/dagster/dagster_tests/definitions_tests/asset_reconciliation_tests/basic_scenarios.py
@@ -1,4 +1,4 @@
-from dagster._core.definitions.asset_reconciliation_sensor import (
+from dagster._core.definitions.auto_materialize_condition import (
     MissingAutoMaterializeCondition,
     ParentMaterializedAutoMaterializeCondition,
 )
